+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.4.2
-* Fix the ActiveRecord Adapter to use upserts instead of find or creates.
+* Update the ActiveRecord Adapter to use upserts instead of first or creates for Assignments.
 
 ## 1.4.1
 * Standardized the return value of batch_pluck.